convert vvvote election schema to a dataclass-based schema
Currently, dataclasses-json is used for serializing dataclasses to JSON.
Deserializing doesn't work well because inheritance is used for Scheme.
Deserializing is not important at the moment, but would be a nice
feature. Maybe we need some kind of type hints in the JSON?
Also, investigate how pydantic would handle the schema.