Create drafts from documents
Documents are displayed with their section headers as links to the draft
creation form. Drafts are hidden until admins push the draft to an
external system or set the visibility property manually.
Changesets link drafts to the changed section of the document.
CustomizableText can be used to include a language-dependent snippet of text
somewhere that should be customizable by admins in the UI.

Closes #42.