Fix/clean up permission rules that check for logged-in users
Morepath allows checking for the type of the current identity with a
predicate. The default is morepath.Identity which always matched whether
the user was logged in or not.
Changing NoIdentity to not be a subclass of morepath.Identity and using
an instance of it instead of the class itself makes it possible to use
the predicate properly.

We already had a fallback permission rule with `has_write_permission_not_logged_in`
which used the predicate, but it had no effect before.