More test stabilization, clean up html asserts
`"content" in res` replaces manual decoding of the body and asserting
things about the result.
Also has the advantage that whitespace is normalized by webtest.

Quoting in HTML causes problems when mimesis generates strings with
quotes, for example. Asserting with the help of beautifulsoup doesn't
have that problem.

Tests look stable now :)