Github action: split up test run in a setup and test part
We now have proper error reporting when something in the test or the
setup fails and the output is split up. Setup loads a lot of
Nix stuff which distracted from the actual test output.

run-tests.sh is not used anymore but can be used to test locally like in
CI. It now accepts an argument to set the location of the Postgres
socket. By default, the created temp dir is used for that.