create_test_db: use test config by default
To create a test DB, the script should work without a config file.
It now imports the test config which means that
EKKLESIA_PORTAL_TEST_DB_URL can be used to set the db connection
without supplying a full config.
We don't have testconfig.yml anymore, so that was quite annoying.

When using the cmdline option `-c`, the script behaves as before
and can be used to create a development database with some default
content.