Proposition: proper handling fo action permissions and "zero" settings
The qualification quorum can be zero (= no supporters needed for
qualification) so the progress bar doesn't make sense in this case. Show
only the number of supporters instead.

Secret voting quorum can also be zero which is interpreted as "requesting secret voting
isn't supported". Don't show the action at all in this case.

If supporting isn't possible only for the current user, just show the data.