Nix: fix poetry file collision in python env
serve_app.nix wouldn't build because of a file collision between poetry
and poetry-core which both provide a poetry package. This was caused by
the overrides from deps.nix that added poetry to propagatedBuildInputs
which propagates poetry to the final python env. Poetry isn't needed
there so we can use buildInputs instead and avoid the collision.