Proposition: show department and area as badges
Department and area are displayed as badges now, like special kinds of tags.

Clicking on a department filters by department and clicking on the area
filters by department and subject area (area is always bound to a
department).

Clicking on a tag now behaves differently: it also filters by department
and subject area as this will be the most common use case: see other
propositions about the same topic, in the same subject area.
Searching for topics globally is generally less helpful but still possible by
removing the department filter with another click.