Fix unsafe repo check in git_version.nix
Newer Git versions check if a dir is owned by the user executing
the Git command. This is not the case if we run a Git command
on the Nix store, so Git refuses to do its work. Add a config
file for Git to fix this.

Fixes static file and docker image building.