Add config arg to run_dev/debug_dev and use 8080
Adding the config file to the default run commands works
better with the documented quick-start setup for dev environments.
If no config is given on the command line, the env var must be
set properly which isn't mentioned directly in the docs. Falling
back to the default config often leads to database connection errors
which are annoying for beginners. Maybe we should always require
a config?

8080 is better supported in browsers. Firefox, for example
denies access to 10080 on localhost by default.